requirements.txt
streamlit
faiss-cpu
sentence-transformers
transformers
numpy